feat: add form to register new user account

+++ src/main/java/ch/finecloud/babytracker/data/service/BabyTrackerService.java
@@ -2,24 +2,31 @@ package ch.finecloud.babytracker.data.service;
 
 import ch.finecloud.babytracker.data.entity.Event;
 import ch.finecloud.babytracker.data.entity.Baby;
+import ch.finecloud.babytracker.data.entity.Role;
+import ch.finecloud.babytracker.data.entity.UserAccount;
 import ch.finecloud.babytracker.data.repository.BabyRepository;
 import ch.finecloud.babytracker.data.repository.EventRepository;
+import ch.finecloud.babytracker.data.repository.UserAccountRepository;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
-@Service 
+@Service
 public class BabyTrackerService {
 
     private final EventRepository eventRepository;
     private final BabyRepository babyRepository;
+    private final UserAccountRepository userAccountRepository;
 
     public BabyTrackerService(EventRepository eventRepository,
-                              BabyRepository babyRepository) {
+                              BabyRepository babyRepository, UserAccountRepository userAccountRepository) {
         this.eventRepository = eventRepository;
         this.babyRepository = babyRepository;
+        this.userAccountRepository = userAccountRepository;
     }
 
 //    @PreAuthorize("hasRole('ADMIN')")
@@ -44,10 +51,16 @@ public class BabyTrackerService {
         return eventRepository.count();
     }
 
-    public void deleteEvent(Event event) {
-        eventRepository.delete(event);
+    @PreAuthorize("hasRole('USER')")
+    public void deleteEvent(Event event, String email) {
+        if (Objects.equals(event.getBaby().getUserAccount().getEmail(), email)) {
+            eventRepository.delete(event);
+        } else {
+            System.err.println("UserAccount " + email + "is not allowed to delete this event.");
+        }
     }
 
+    @PreAuthorize("hasRole('USER')")
     public void saveEvent(Event event) {
         if (event == null) {
             System.err.println("Event is null. Are you sure you have connected your form to the application?");
@@ -73,15 +86,42 @@ public class BabyTrackerService {
         }
     }
 
-    public void saveBaby(Baby baby) {
+    @PreAuthorize("hasRole('USER')")
+    public void saveBaby(Baby baby, String email) {
         if (baby == null) {
             System.err.println("Baby is null. Are you sure you have connected your form to the application?");
+            return;
+        }
+        Optional<UserAccount> userAccount = userAccountRepository.findUserAccountByEmail(email);
+        if (userAccount.isPresent()) {
+            baby.setUserAccount(userAccount.get());
+        } else {
+            System.err.println("UserAccount is unknown. Are you sure you have connected your form to the application?");
             return;
         }
         babyRepository.save(baby);
     }
 
-    public void deleteBaby(Baby baby) {
-        babyRepository.delete(baby);
+    @PreAuthorize("hasRole('USER')")
+    public void deleteBaby(Baby baby, String email) {
+        if (Objects.equals(baby.getUserAccount().getEmail(), email)) {
+            babyRepository.delete(baby);
+        } else {
+            System.err.println("UserAccount " + email + "is not allowed to delete this baby.");
+        }
+    }
+
+    public String addUserAccount(String email, String password) {
+        UserAccount userAccount = new UserAccount();
+        userAccount.setEmail(email);
+        userAccount.setPassword(password);
+        userAccount.setRole(Role.USER);
+        UserAccount savedUserAccount;
+        try {
+            savedUserAccount = userAccountRepository.save(userAccount);
+        } catch (Exception e) {
+            return "Error: " + e.getMessage();
+        }
+        return savedUserAccount.getEmail();
     }
 }
